Added tabs to the app
